rename core.py -> accessor.py
