+++ src/xoak/__init__.py
@@ -2,4 +2,4 @@ __version__ = "0.1.0"
 
 
 from .index import IndexAdapter, indexes, register_index
-from .core import XoakAccessor
+from .accessor import XoakAccessor
